Rename packages:  - "network" to "api"  - "repository" to "domain" Move package "entities" into domain
