+++ app/src/main/java/ru/aevd/androidacademymovieapp/api/RetrofitModule.kt
@@ -1,4 +1,4 @@
-package ru.aevd.androidacademymovieapp.network
+package ru.aevd.androidacademymovieapp.api
 
 import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
 import kotlinx.serialization.json.Json
