+++ app/src/main/java/ru/aevd/androidacademymovieapp/domain/MoviePersistent.kt
@@ -1,7 +1,7 @@
-package ru.aevd.androidacademymovieapp.repository
+package ru.aevd.androidacademymovieapp.domain
 
 import android.content.Context
-import ru.aevd.androidacademymovieapp.entities.Movie
+import ru.aevd.androidacademymovieapp.domain.entities.Movie
 
 class MoviePersistent(private val context: Context) {
     suspend fun loadMovies(): List<Movie> = loadMovies(context)
