+++ app/src/main/java/ru/aevd/androidacademymovieapp/ui/adapters/OnMoviesItemClicked.kt
@@ -1,6 +1,6 @@
 package ru.aevd.androidacademymovieapp.ui.adapters
 
-import ru.aevd.androidacademymovieapp.entities.Movie
+import ru.aevd.androidacademymovieapp.domain.entities.Movie
 
 interface OnMoviesItemClicked {
     fun onClick(movie: Movie)
