+++ app/src/main/java/ru/aevd/androidacademymovieapp/ui/viewmodels/MoviesDetailsViewModel.kt
@@ -1,8 +1,8 @@
-package ru.aevd.androidacademymovieapp.viewmodels
+package ru.aevd.androidacademymovieapp.ui.viewmodels
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import ru.aevd.androidacademymovieapp.entities.Movie
+import ru.aevd.androidacademymovieapp.domain.entities.Movie
 
 class MoviesDetailsViewModel(movie: Movie): ViewModel() {
     private val _movie = MutableLiveData(movie)
